==> old_syntax/00_Data Preparation #2-GHSL_2010_Data-Clipping.ipynb <==
import geopandas as gpd
import pandas as pd
import numpy as np
import glob
import rasterio
from shapely.geometry import Polygon
from rasterio.mask import mask
from rasterio.merge import merge

from google.cloud import storage

client = storage.Client()
ls_=list(client.list_blobs('bps-gcp-bucket', prefix='Degree of Urbanisation/GHSL data/2010/zip/'))

storage_client = storage.Client()
for i in ls_:
    data_=str(i).split('/')[4].split(',')[0]
    if len(data_)>3:
        bucket = storage_client.get_bucket('bps-gcp-bucket')
        source_blob_name='Degree of Urbanisation/GHSL data/2010/zip/'+data_
        blob = bucket.blob(source_blob_name)
        blob.download_to_filename('data/input/GHSL Data/2010/zip/'+data_)

import zipfile
data_=glob.glob('data/input/GHSL Data/2010/zip/*.zip')
for i in data_:
    with zipfile.ZipFile(i, 'r') as zip_ref:
        zip_ref.extractall('data/input/GHSL Data/2010/tif')

data_prov=gpd.read_file('data/input/Master_Data/Prov_20201_ByIndo(1).shp').to_crs('ESRI:54009')[['geometry','kdprov']]

ls_j=glob.glob('data/input/GHSL Data/2010/tif/*.tif')
for j in ls_j:
    with rasterio.open(j) as src:
        xmin=src.bounds.left
        xmax=src.bounds.right
        ymin=src.bounds.bottom
        ymax=src.bounds.top
        gpd_box=gpd.GeoDataFrame({'geometry':[Polygon([(xmin,ymin), (xmax, ymin), (xmax, ymax), (xmin, ymax)])]},crs='ESRI:54009')
        y_=gpd.sjoin(gpd_box,data_prov,op='intersects')['kdprov'].tolist()
        name_=j.split('/')[-1]
        for y1_ in y_:
            shapes=data_prov.query('kdprov==@y1_').geometry
            out_image, out_transform = mask(src, shapes, crop=True)
            out_meta = src.meta
            out_meta.update({"driver": "GTiff",
                 "height": out_image.shape[1],
                 "width": out_image.shape[2],
                 "transform": out_transform})
            with rasterio.open("data/input/GHSL Data/2010/tif_prov/2010_"+y1_+"_"+name_, "w", **out_meta) as dest:
                dest.write(out_image)

for i in data_prov.kdprov:
    y_=glob.glob('data/input/GHSL Data/2010/tif_prov/2010_'+i+'*')
    dest, output_transform=merge(y_)
    with rasterio.open(y_[0]) as src:
        out_meta = src.meta.copy()    
    out_meta.update({"driver": "GTiff",
                 "height": dest.shape[1],
                 "width": dest.shape[2],
                 "transform": output_transform})
    with rasterio.open("data/input/GHSL Data/2010/prov/2010_"+i+".tif", "w", **out_meta) as dest1:
        dest1.write(dest)

def upload_blob(bucket_name, source_file_name, destination_blob_name):
    """Uploads a file to the bucket."""
    storage_client = storage.Client()
    bucket = storage_client.bucket(bucket_name)
    blob = bucket.blob(destination_blob_name)

    blob.upload_from_filename(source_file_name)

    print(
        f"File {source_file_name} uploaded to {destination_blob_name}."
    )

upload_2010=glob.glob('data/input/GHSL Data/2010/prov/2010_*')
for i in upload_2010:
    y_=i.split('/')[-1]
    upload_blob('bps-gcp-bucket',i,'Degree of Urbanisation/GHSL data/2010/prov/'+y_)
