==> old_syntax/01_PopDens.ipynb <==
import pandas as pd
from dbfread import DBF

data_ruta=pd.DataFrame(iter(DBF('rekap_sp2020.dbf')))

import geopandas as gpd
from glob import glob

ls_sls=glob('sls_*.gpkg')

ls_sls

data_sls=gpd.GeoDataFrame()
for i in ls_sls:
    data_sls_i=gpd.read_file(i)
    data_sls=data_sls_i.append(data_sls)
    data_sls_i=None

data_ruta.head(5)

data_ruta['idsls']=data_ruta.KODE_SLS.apply(lambda y: str(y)[:14])

#calculate per sls
data_ruta_compiled=pd.pivot_table(data_ruta,index='idsls',values='TOTAL_PEND',aggfunc='sum').reset_index()

data_sls['idsls']=data_sls.idsls.astype(str)
data_peta_w_ruta=data_sls[['geometry','idsls']].merge(data_ruta_compiled,how='left').fillna(0)

data_peta_w_ruta['area_pop']=data_peta_w_ruta.to_crs('ESRI:53034').area

#open built-up
data_builtup=gpd.read_file('BUILTUP_515253_REV.gpkg').to_crs(4326).dissolve()
data_builtup=data_builtup[['geometry']]

data_builtupsls=gpd.overlay(data_builtup,data_peta_w_ruta,
                            how='intersection').explode().reset_index()[['idsls','TOTAL_PEND','area_pop','geometry']]

data_builtupsls['area_built']=data_builtupsls.to_crs('ESRI:53034').area

table_area=pd.pivot_table(data_builtupsls[['idsls','area_built']],
                          index='idsls',values='area_built',aggfunc='sum').reset_index()

table_area.columns=['idsls','area_built_total']

data_builtupsls=data_builtupsls.merge(table_area).assign(per_built=lambda y:y.area_built/y.area_built_total*100 )
data_builtupsls['pop_built']=data_builtupsls.per_built*data_builtupsls.TOTAL_PEND/100

data_builtupsls.to_file('sls_builtup_5000.gpkg',driver='GPKG')

#generate S2GEN 1km 

from s2gen import s2gen
y_=data_builtupsls.total_bounds
kml = s2gen(y_[3],y_[1],y_[2],y_[0], 13)

with open("5000.kml", "w") as f:
    f.write(kml)

import fiona

gpd.io.file.fiona.drvsupport.supported_drivers['KML'] = 'rw'
grid_1km=gpd.read_file('5000.kml',driver='KML')

data_builtupslss2=gpd.overlay(grid_1km,data_builtupsls,how='intersection')

data_builtupslss2['area_built_after']=data_builtupslss2.to_crs('ESRI:53034').area

data_builtupslss2['per_built_after']=data_builtupslss2.area_built_after/data_builtupslss2.area_built

data_builtupslss2['pop_built_s2']=data_builtupslss2.pop_built*data_builtupslss2.per_built_after

tabel_pop_built_s2=pd.pivot_table(data_builtupslss2,index='Name',
                                  values=['pop_built_s2','per_built_after'],aggfunc='sum').reset_index()

grid_with_pop=grid_1km.merge(tabel_pop_built_s2,how='left').fillna(0)
data_gridpops2=gpd.sjoin(grid_with_pop,data_sls[['geometry']],how='inner',op='intersects')


(data_gridpops2[['Name','geometry','pop_built_s2','per_built_after']]
 .dissolve(by='Name').to_file('pop_dens_s2_5000.gpkg',driver='GPKG'))
