==> old_syntax/02_DEGURBA #2 Deliniating Small Spatial Unit.ipynb <==
import geopandas as gpd
import pandas as pd

data_pop=gpd.read_file('data/output/1_population_density/grid_pop_19_100m.gpkg')
data_degurba_grod=gpd.read_file('data/output/2_grid_degurba/degurba_1900.gpkg')

data_desa_raster=gpd.read_file('data/output/0_raster_data_desa/desa_pend_rasterize_19.gpkg')

pop_w_grid=data_pop.sjoin(data_degurba_grod)[['j_pend','p_builtup','classification','geometry']]

data_desa=data_desa_raster.sjoin(pop_w_grid)

data_desa

tab_data=pd.pivot_table(data_desa,index='iddesa',columns='classification',values='j_pend',aggfunc='sum').fillna(0).reset_index()

def define_LAU(y):
    total_pop=y.CENTRE+y.CLUSTER+y.RURAL
    if total_pop==0:
        return 'RURALS'
    elif (y.CENTRE/total_pop)>=0.5:
        return 'CITIES'
    elif (((y.CENTRE/total_pop)<0.5)&((y.RURAL/total_pop)<0.5)):
        return 'TOWNS'
    else:
        return 'RURALS'

tab_data['LAU']=tab_data.apply(lambda y:define_LAU(y),axis=1)

tab_data

kdprov='19'
peta_desa=gpd.read_file(r"data/input/Master_Data/DESA_20201_BYINDO.gpkg").query('kdprov==@kdprov')

peta_desa_LAU=peta_desa.merge(tab_data[['iddesa','LAU']],how='left').fillna('RURALS')

peta_desa_LAU.to_file('data/output/3_small_spatial_unit_degurba/small_spatial_unit_'+kdprov+'.gpkg',driver='GPKG')

from dbfread import DBF
mfd_=pd.DataFrame(iter(DBF('data/input/Master_Data/mfd_2020_1.dbf')))

mfd_['iddesa']=mfd_.PROP+mfd_.KAB+mfd_.KEC+mfd_.DESA

peta_desa_LAU.merge(mfd_[['iddesa','DRH']]).to_file('data/output/3_small_spatial_unit_degurba/small_spatial_unit_'+kdprov+'.gpkg',driver='GPKG')

peta_desa_LAU.merge(mfd_[['iddesa','DRH']]).pivot_table(index='DRH',columns='LAU',values='iddesa',aggfunc='count')

peta_desa_LAU.merge(mfd_[['iddesa','DRH']]).query('LAU==\'CITIES\'').query('DRH==\'2\'')
