==> old_syntax/03_LAU calculation.ipynb <==
import geopandas as gpd
import pandas as pd

data_pop_densls=gpd.read_file('sls_builtup_5300.gpkg')
data_degurba=gpd.read_file('degurba_5300_rev_w_buitup_adjustment_3.gpkg')

data_pop_dens_degurba=gpd.overlay(data_degurba,data_pop_densls,how='intersection')

data_pop_dens_degurba.head(2)

data_pop_dens_degurba['iddesa']=data_pop_dens_degurba.idsls.apply(lambda y: str(y)[:10])

tab_classi=pd.pivot_table(data_pop_dens_degurba,index='iddesa',columns='classification',
                          values='pop_built',aggfunc='sum').fillna(0).reset_index()

def define_LAU(y):
    total_pop=y.CENTRE+y.CLUSTER+y.RURAL
    if total_pop==0:
        return 'RURALS'
    elif (y.CENTRE/total_pop)>=0.5:
        return 'CITIES'
    elif (((y.CENTRE/total_pop)<0.5)&((y.RURAL/total_pop)<0.5)):
        return 'TOWNS'
    else:
        return 'RURALS'

tab_classi['LAU']=tab_classi.apply(lambda y:define_LAU(y),axis=1)

peta_desa=gpd.read_file('DESA_20201_BYINDO.gpkg')

from dbfread import DBF

mfd_2020_1=pd.DataFrame(iter(DBF('mfd_2020_1.dbf')))

peta_desa_5100=peta_desa.query('kdprov==\'53\'')

peta_desa_5100=peta_desa_5100.merge(tab_classi[['iddesa','LAU']],how='left').fillna('RURALS')

mfd_desa_5100_UR=mfd_2020_1.query('PROP==\'53\'').assign(iddesa=lambda y:y.PROP+y.KAB+y.KEC+y.DESA)[['iddesa','DRH']]

peta_desa_5100.merge(mfd_desa_5100_UR,how='left').rename(
    columns={'DRH':'UR'}).fillna({'UR':2}).to_file('desa_5300_LAU_UR_rev2.gpkg',driver='GPKG')

peta_desa_5100.merge(mfd_desa_5100_UR,how='left').rename(
    columns={'DRH':'UR'}).fillna({'UR':2}).pivot_table(columns='LAU',index='UR',values='iddesa',aggfunc='count')

peta_desa_5100.merge(mfd_desa_5100_UR,how='left').rename(
    columns={'DRH':'UR'}).fillna({'UR':2}).query('(UR=="1") and (LAU=="RURALS")')
